feat(modularization): S-3 extract router/nav core to web/modules/app-core.js
Extracts switchPage(), renderNav(), renderStaticI18n(), and boot() from
the 16k-line app.js monolith into a dedicated ES module.

Design decision -- registration pattern over direct dispatch:
  registerPage(name, { enter?, leave? }) decouples the router from page-
  specific load functions. The leave hook can abort navigation (settings
  dirty guard). registerAlias(from, to, before?) handles the team->agents
  redirect with its _agentsActiveView side-effect staying in app.js.

What moved to app-core.js:
  - switchPage() with alias resolution + leave/enter lifecycle hooks
  - NAV_I18N + renderNav()
  - PAGE_HEADER_I18N + renderStaticI18n()
  - boot(): nav click binding, sidebar toggle, hashchange listener,
    i18n initial render + onLangChange re-render, routeFromHash()

What stays in app.js:
  - All page load/poll functions (loadOverview, startActivityPoll, ...)
  - 24 registerPage() calls + 1 registerAlias() call
  - DOMContentLoaded -> boot()

Removed from app.js:
  - navLinks/pages DOM queries (~2 lines)
  - confirmSettingsLeave() (now the settings leave hook)
  - switchPage() function (~43 lines)
  - Sidebar toggle code (~10 lines)
  - Nav click listeners (~10 lines)
  - NAV_I18N + renderNav() (~20 lines)
  - PAGE_HEADER_I18N + renderStaticI18n() + static i18n maps (~140 lines)
  - DOMContentLoaded boot block (~10 lines)
  - Hash routing IIFE (~11 lines)
  - onLangChange re-render block (~8 lines, now in app-core.js boot())

Test fixes: federation-ui-contract and approvals-ui-contract now read
app-core.js for nav key / page-header assertions; router dispatch checks
updated from 'pageId === X' to 'registerPage(X)'.

HTTP verified: /modules/app-core.js -> 200 application/javascript.
All 2868 tests pass; tsc --noEmit clean; node --check clean.
